Personal Learning/04_image_search/pyproject.toml
[project]
name = "04-image-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.41.0",
    "openai-agents>=0.3.3",
    "pillow>=11.3.0",
    "pytesseract>=0.3.13",
    "tavily-python>=0.7.12",
]
